% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAnnot.R
\name{addAnnot}
\alias{addAnnot}
\title{Add annotation to a given object through the attribute 'gen.info'}
\usage{
addAnnot(x)
}
\arguments{
\item{x}{any object (usually 'parGADA' object)}
}
\value{
the same object with an extra attribute called 'gen.info' including the annotation
}
\description{
Add annotation to a given object through the attribute 'gen.info'
}
