\name{cat_sum1}
\alias{cat_sum1}
\title{Intermediate level function}
\description{Intermediate level function for categorical moderation. sums relevant es information by each moderator level. 
}
\usage{
cat_sum1(meta, mod)
}
\value{ Fixed effects means and weights for each level of moderator. 
}
\references{ Cooper, H., Hedges, L.V., & Valentine, J.C. (2009). The handbook of research synthesis and meta-analysis (2nd edition). New York: Russell Sage Foundation. 
}
\author{ AC Del Re & William T. Hoyt
}
\keyword{internal}

