\name{dat.delre}
\alias{dat.delre}
\title{Subset of simulated psychotherapy treatment studies}
\description{This data set includes 10 randomly selected studies from a population of simulated psychotherapy treatment studies. The population standardized mean difference for outcome one is d=0.50 and for outcome two g=0.80. The data set also includes two correlated study-level variables (i.e., moderators): treatment duration (\code{dose}) and average participant distress (\code{stress}). This data comes from: Del Re, A. C. & Fluckiger (In prep). Meta-analysis. \emph{Handbook of Clinical Psychology}.}
\keyword{data}