\name{dat.delre1}
\alias{dat.delre1}
\title{Simulated psychotherapy treatment studies for outcome one}
\description{This data set includes 1000 simulated psychotherapy treatment studies. The population standardized mean difference for outcome one is d=0.50. The data set also includes two correlated study-level variables (i.e., moderators): treatment duration (\code{dose}) and average participant distress (\code{stress}). This data comes from: Del Re, A. C. & Fluckiger (In prep). Meta-analysis. \emph{Handbook of Clinical Psychology}.}
\keyword{data}