\name{dat.hoyt}
\alias{dat.hoyt}
\title{Subset of simulated data to demonstrate aggregation for dependent effect sizes}
\description{Subset of simulated data to demonstrate aggregation for dependent effect sizes. This data comes from: Hoyt, W. T., & Del Re, A. C. (2013).  Comparison of methods for aggregating dependent effect sizes in meta-analysis. Manuscript submitted for publication.}
\keyword{data}