\name{dat.sim.es}
\alias{dat.sim.es}
\title{Subset of simulated psychotherapy treatment studies (k=8) with dependent effect sizes}
\description{This data set includes effect sizes for treatment vs control groups for a random subset of 8   simulated psychotherapy treatment studies (total population k = 1000). The known population standardized mean difference for outcome one is g = 0.50 and for outcome two g = 0.80. The data set also includes two correlated study-level variables (i.e., moderators): treatment duration (\code{dose}) and average participant distress (\code{stress}). 

This data accompanies a tutorial for meta-analysis: 

Del Re, A. C. (2015). A Practical Tutorial on Conducting Meta-Analysis in R. \emph{The Quantitative Methods for Psychology}.}
\keyword{data}