\name{dat.sim.raw}
\alias{dat.sim.raw}
\title{Subset of simulated psychotherapy treatment studies (k=8) with means and SD}
\description{This data set includes means and standard deviations for treatment and control groups for a subset of k = 8 simulated psychotherapy treatment studies from a population of k = 1000 (see dat.sim1 and dat.sim2 for the k = 1000 studies for outcome 1 and 2, respectfully). Although effect sizes are not calculated in this dataset, the known population standardized mean difference for outcome one is d=0.50 and for outcome two g=0.80. The data set also includes two correlated study-level variables (i.e., moderators): treatment duration (\code{dose}) and average participant distress (\code{stress}). 

This data accompanies a tutorial for meta-analysis: 

Del Re, A. C. (2015). A Practical Tutorial on Conducting Meta-Analysis in R. \emph{The Quantitative Methods for Psychology}.}
\keyword{data}