\name{dat.sim2}
\alias{dat.sim2}
\title{Simulated psychotherapy treatment studies for 'outcome two'}
\description{This data set includes 1000 simulated psychotherapy treatment studies. The population standardized mean difference for outcome two is d=0.80. The data set also includes two correlated study-level variables (i.e., moderators): treatment duration (\code{dose}) and average participant distress (\code{stress}). 

This data accompanies a tutorial for meta-analysis: 

Del Re, A. C. (2015). A Practical Tutorial on Conducting Meta-Analysis in R. \emph{The Quantitative Methods for Psychology}.}
\keyword{data}