% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMosaics.R
\name{getMosaics}
\alias{getMosaics}
\title{Get mosaic alterations as a GenomicRange object}
\usage{
getMosaics(x, ...)
}
\arguments{
\item{x}{an object from 'parBE.B.deviation' function}
}
\value{
a 'GenomicRange' object having mosaic alterations
}
\description{
Get mosaic alterations as a GenomicRange object
}
