% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a_difference.R
\name{make_a_difference}
\alias{make_a_difference}
\title{Make a Difference Function}
\usage{
make_a_difference(data, col1, col2)
}
\arguments{
\item{data}{The dataset containing the chosen variables to apply this function on}

\item{col1}{The first chosen column/variable within the dataset.  It must contain only numeric values.}

\item{col2}{The second chosen column/variable within the dataset. It must contain only numeric values.
OUTPUTS:}
}
\value{
The creation of a "Difference" column in the dataset containing the calculated differences of all the data points between the 2 variables.
}
\description{
Introducing the 'Make a Difference' Function - This function calculates the difference between 2 variables, then creates a new column to hold these calculations in the dataset.
INPUTS:
}
