% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parBE.B.deviation.R
\name{parBE.B.deviation}
\alias{parBE.B.deviation}
\title{Parallel backward elimination of B-deviation}
\usage{
parBE.B.deviation(x, Samples, T, MinSegLen, verbose = TRUE, mc.cores = 1,
  ...)
}
\arguments{
\item{x}{an object of class 'parGADA' obtained using 'setupParGADA.B.deviation' function}

\item{Samples}{the samples to be analyzed (obtained from 'x' by default)}

\item{T}{critical value of the backward elimination algorithm (see 'BackwardElimination' function)}

\item{MinSegLen}{minimum number of probes each segment must contain (see 'BackwardElimination' function)}

\item{verbose}{is information about each individual showed in the console? Default is TRUE}

\item{mc.cores}{number of cores to be used when using multiple cores (see argument 'mc.cores' from 'mclapply' function of parallel library)}

\item{...}{other arguments passed through the function 'BackwardElimination'}
}
\value{
a 'GenomicRange' object having mosaic alterations
}
\description{
Parallel backward elimination of B-deviation
}
