% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQMosaic.R
\name{plotQMosaic}
\alias{plotQMosaic}
\title{Plot LRR and BAF of a given chromosome}
\usage{
plotQMosaic(x, chr, sample, regions, delim = NULL, title = TRUE,
  col.dots = c("black", "red"), ...)
}
\arguments{
\item{x}{an object of class 'parGADA'}

\item{chr}{chromosome of interest}

\item{sample}{sample of interest}

\item{regions}{regions to be highlighted}

\item{delim}{same as xlim. Default is NULL}

\item{title}{should title be added? Default is TRUE}
}
\value{
a plot highlighting the altered regions
}
\description{
Plot LRR and BAF of a given chromosome
}
