% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZoomQMosaic.R
\name{plotZoomQMosaic}
\alias{plotZoomQMosaic}
\title{Zoom plot of LRR and BAF on regions of interest}
\usage{
plotZoomQMosaic(x, chr, sample, regions, delim, ...)
}
\arguments{
\item{x}{an object of class 'parGADA'}

\item{chr}{chromosome of interest}

\item{sample}{sample of interest}

\item{regions}{regions to be highlighted}

\item{delim}{same as xlim. Default is NULL}
}
\value{
a pannel including two plots
}
\description{
Zoom plot of LRR and BAF on regions of interest
}
