% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupGADA.B.deviation.R
\name{setupGADA.B.deviation}
\alias{setupGADA.B.deviation}
\title{Imports array data to MAD}
\usage{
setupGADA.B.deviation(file, NumCols, GenoCol, log2ratioCol, BAFcol,
  name.geno = c("AA", "AB", "BB"), chrs = c(1:22, "X", "Y"),
  MarkerIdCol = 1, ChrNameCol = 2, ChrPosCol = 3, sort = TRUE,
  orderProbes, sep = "\\t", saveGenInfo = TRUE)
}
\arguments{
\item{file}{A pennCNV-style file}

\item{NumCols}{The number of columns in 'file'}

\item{GenoCol}{The column in 'file' containing genotypes}

\item{log2ratioCol}{The column in 'file' containing log2ratio intensities}

\item{name.geno}{the name of the genotypes (default "AA", "AB" and "BB)}

\item{chrs}{The names of the chromosomes. Default 1, ..., 22, X, Y}

\item{MarkerIdCol}{The column in 'file' containing the name of the marker. Default first column}

\item{ChrNameCol}{The column in 'file' containing the chromosome. Default second column}

\item{ChrPosCol}{The column in 'file' containing the genomic position. Default third column}

\item{sort}{should data be sorted by genomic position? The default is TRUE.}

\item{orderProbes}{used to avoid sorting}

\item{sep}{how data columns are separated. Default by tab}

\item{saveGenInfo}{Should annotation data be saved? The default is TRUE. See details}

\item{BAFCol}{The column in 'file' containing B-allele frequency (BAF)}
}
\description{
Imports array data to MAD
}
