% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupParGADA.B.deviation.R
\name{setupParGADA.B.deviation}
\alias{setupParGADA.B.deviation}
\title{A parallel version of setupGADA.B.deviation function}
\usage{
setupParGADA.B.deviation(folder, files, chrs = c(as.character(1:22), "X",
  "Y"), verbose = TRUE, sort = TRUE, MarkerIdCol = 1, ChrNameCol = 2,
  ChrPosCol = 3, mc.cores = 1, ...)
}
\arguments{
\item{folder}{The folder where data is stored. Not required if the working directory contains a 'rawData' folder}

\item{files}{The names of the files with pennCNV-format fiels. Not required. By default all files in the 'rawData' folder are analyzed}

\item{chrs}{The names of the chromosomes. Default 1, ..., 22, X, Y}

\item{verbose}{Should information about process be printed in the console? The default is TRUE}

\item{sort}{Should data be sorted by genomic position? Default is TRUE}

\item{MarkerIdCol}{The column in 'file' containing the name of the marker. Default first column}

\item{ChrNameCol}{The column in 'file' containing the chromosome. Default second column}

\item{ChrPosCol}{The column in 'file' containing the genomic position. Default third column}

\item{mc.cores}{number of cores to be used when using multiple cores (see argument 'mc.cores' from 'mclapply' function of 'parallel' library)}

\item{...}{Other arguments passed through 'setupGADA'}
}
\description{
A parallel version of setupGADA.B.deviation function
}
