\name{mad-package}
\alias{mad-package}
\alias{mad}
\docType{package}

\title{
Mosaic Alteration Detector (MAD)
}

\description{
The package implements an efficient algorithm to detect mosaic alterations using SNP array data
}

\details{
\tabular{ll}{
Package: \tab gada\cr
Type: \tab Package\cr
Version: \tab 0.7-3\cr
Date: \tab 2008-11-26\cr
License: \tab GPL \cr
URL: \tab http://www.creal.cat/jrgonzalez/software  \cr
LazyLoad: \tab yes\cr
}
}

\author{
Juan R Gonzalez <jrgonzalez@creal.cat>

Maintainer: Juan R Gonzalez <jrgonzalez@creal.cat>

}


\references{
JR Gonzalez, B Rodriguez-Santiago, A Caceres, R Pique-Regi, N Rothman, SJ Chanock, 
L Armengol, LA Perez-Jurado. A fast and accurate method to detect allelic genomic imbalances underlying mosaic rearrangements using SNP array data
BMC Bioinformatics 2011, 12:166.
}

\keyword{ package }



